requirements.txt
django
djangorestframework
cx_oracle